requirements.txt
Flask==1.1.1
firebase_admin
Blueprint
pylint
flask-restx
google-cloud-bigquery